requirements.txt
pandas
numpy
flask
flask-cors
transformers
torch
torchvision
Pillow